fix: smooth two-finger scroll and add Windows tray icon
- Replace pyautogui.scroll() (discrete clicks) with ctypes MOUSEEVENTF_WHEEL
  accumulator — backend fires one notch per 25px of finger movement, making
  scroll proportional and non-jerky
- Send raw float dy from frontend instead of Math.round(), giving the
  accumulator sub-pixel precision
- Add 150ms cooldown after second finger lifts before single-finger move
  resumes, preventing cursor jump that triggered Windows auto-scroll mode
- Add tray.ps1: native Windows Forms tray icon (green/red status dot,
  double-click opens browser, right-click start/stop/exit)
- Add tray.py: pystray fallback (unused, kept for reference)

+++ agents/laptop/ws_input.py
@@ -22,7 +22,7 @@ def attach(app: FastAPI) -> None:
                     if t == "move":
                         input_win.move_rel(float(frame["dx"]), float(frame["dy"]))
                     elif t == "scroll":
-                        input_win.scroll(int(frame["dx"]), int(frame["dy"]))
+                        input_win.scroll(float(frame["dx"]), float(frame["dy"]))
                     elif t == "down":
                         input_win.mouse_down(str(frame["btn"]))
                     elif t == "up":
